backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Requests
httpx==0.26.0
requests==2.31.0

# Redis Cache
redis==5.0.1
hiredis==2.3.2

# LLM APIs
anthropic==0.18.1
google-generativeai==0.3.2

# Database (optional for future)
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment
python-dotenv==1.0.0

# CORS
python-cors==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
